DuplicatorScript for fast fences, Spawnpoint text visible toggle, TransformLooper for..effects, updated prefabs
Apply TransformDuplicator to any game object, give it the values you
want to duplicate an prefab with and a prefab to duplicate, press the Do
button and your dupes appear. TransformLooper makes a runtime
conveyorbelt/stream effect. ShowTextMeshInEditor toggles the meshtext to
empty at runtime and back to a desired text during Editing mode.